+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to
 
 - Diagnostic Settings entities and relationships for Network Security Groups
 
+### Fixed
+
+- Fixed the way `IntegrationProviderAPIError` exposes error code/message of
+  `node-fetch` errors, such as `ECONNRESET`.
+
 ## 5.9.0 - 2020-12-17
 
 ### Added